cut excessive tests
